@@ -1,5 +1,5 @@
 {
-  "name": "vite-reactts-eslint-prettier-starter",
+  "name": "zmanim",
   "version": "0.0.1",
   "scripts": {
     "dev": "vite",
@@ -12,60 +12,59 @@
     "type-check": "tsc"
   },
   "dependencies": {
-    "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.8",
-    "@commitlint/cli": "^16.2.4",
-    "@commitlint/config-conventional": "^16.2.4",
-    "@emotion/react": "^11",
-    "@emotion/styled": "^11",
+    "@commitlint/cli": "^17.3.0",
+    "@commitlint/config-conventional": "^17.3.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@mapbox/mapbox-gl-geocoder": "^5.0.1",
     "@mapbox/timespace": "^2.0.4",
-    "@types/node": "^17.0.30",
-    "axios": "^0.27.2",
-    "date-fns": "^2.28.0",
+    "@mui/material": "^5.10.17",
+    "@tanstack/react-query": "^4.20.4",
+    "@tanstack/react-query-devtools": "^4.20.4",
+    "axios": "^1.2.2",
+    "date-fns": "^2.29.3",
     "framer-motion": "^6",
-    "i18next": "^21.6.16",
-    "i18next-parser": "^6.4.0",
+    "i18next": "^22.4.7",
+    "i18next-parser": "^7.1.0",
     "kosher-zmanim": "^0.8.2",
     "lodash": "^4.17.21",
-    "luxon": "^2.3.2",
-    "mapbox-gl": "^2.8.2",
-    "react": "^17.0.2",
+    "luxon": "^3.2.0",
+    "mapbox-gl": "^2.11.1",
+    "react": "^18.2.0",
     "react-calendar": "^3.7.0",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.31.1",
-    "react-i18next": "^11.16.7",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.41.3",
+    "react-i18next": "^12.1.1",
     "react-icons": "^4.3.1",
-    "react-map-gl": "^7.0.13",
+    "react-map-gl": "^7.0.20",
     "react-map-gl-geocoder": "^2.2.0",
-    "react-query": "^3.38.0",
-    "react-router-dom": "6",
-    "sass": "^1.51.0"
+    "react-router-dom": "^6.6.1"
   },
   "devDependencies": {
-    "@types/luxon": "^2.3.1",
+    "@types/lodash": "^4.14.191",
+    "@types/luxon": "^3.2.0",
     "@types/mapbox-gl": "^2.7.0",
-    "@types/mapbox__mapbox-gl-geocoder": "^4.7.2",
-    "@types/react": "^17.0.39",
-    "@types/react-calendar": "^3.5.0",
-    "@types/react-dom": "^17.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
+    "@types/mapbox__mapbox-gl-geocoder": "^4.7.3",
+    "@types/node": "^18.11.18",
+    "@types/react-calendar": "^3.9.0",
+    "@types/react-dom": "^18.0.6",
+    "@typescript-eslint/eslint-plugin": "^5.48.0",
+    "@typescript-eslint/parser": "^5.48.0",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "eslint": "^8.14.0",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint": "^8.31.0",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-simple-import-sort": "^8.0.0",
     "husky": "4.3.8",
-    "lint-staged": "^12.4.1",
+    "lint-staged": "^13.1.0",
     "pre-commit": "^1.2.2",
-    "prettier": "^2.6.2",
-    "typescript": "^4.6.4",
-    "vite": "^2.9.6",
-    "vite-tsconfig-paths": "^3.4.1"
+    "prettier": "^2.8.1",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.3",
+    "vite-tsconfig-paths": "^4.0.3"
   },
   "resolutions": {
     "csstype": "3.0.10"
