@@ -1,16 +1,12 @@
-import { Box, Container, SimpleGrid, Stack, useColorModeValue } from '@chakra-ui/react';
+import { Box, Container } from '@mui/material';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
 export const Footer = () => {
   const { t } = useTranslation();
   return (
-    <Box bg={useColorModeValue('gray.50', 'gray.900')} color={useColorModeValue('gray.700', 'gray.200')}>
-      <Container as={Stack} maxW="6xl" py={10}>
-        <SimpleGrid templateColumns={{ sm: '1fr 1fr', md: '2fr 1fr 1fr 1fr 1fr' }} spacing={8}>
-          <Box>{t('footer.title')}</Box>
-        </SimpleGrid>
-      </Container>
+    <Box>
+      <Container />
     </Box>
   );
 };
