@@ -5,17 +5,13 @@ import React, { useState } from 'react';
 
 import { ZMANIM_KEYS } from '../../../constants/common';
 import { useCalendar } from '../../../providers/CalendarProvide';
-import { ZmanimInfoFilters } from './ZmanimInfoFilters';
 
 export const ZmanimInfo = () => {
   const { zmanimJson } = useCalendar();
   const [show, setShow] = useState(ZMANIM_KEYS);
-  const handleFilters = (val: { [key: string]: boolean }) => {
-    setShow(Object.keys(val).filter((k) => val[k]));
-  };
+
   return (
     <div>
-      {/*<ZmanimInfoFilters handleFormSubmit={handleFilters} />*/}
       {map(
         pick(
           zmanimJson?.BasicZmanim,
